nlp_service/requirements.txt
Flask==2.3.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
torch==2.0.0
transformers==4.30.0
spacy==3.5.0
PyPDF2==3.0.1
scikit-learn==1.3.0
pandas==2.0.0
numpy==1.24.0
requests==2.31.0
